Replaced ADS1115 with ADS1015 on PT page
The ADS1115 is being used to read the voltages provided by the Helix
system to the pressure transducers. The ADS1015 is already doing
the same job for the Hall Effect sensors, however the ADS1015 has
4 bits lower accuracy. High accuracy is not needed for this
application so be replacing the ADS1115 with the ADS1015 we can
reduce cost and more importantly reduce BOM size making the project
logistics easier to manage.